[Refactor & Addition] Changes
- Force lua hooks for returning functions to return a value
- Change `aether::lua::manager::try_register_hook(...)` to only accept rvalue for hook callbacks
- Use floats for debug inside renderer instead of doubles
- Pass by value for math lua binding
- Add fmt lua binding

+++ aether/game.cc
@@ -124,7 +124,7 @@ void game::run() {
 		}
 
 #if defined(AETHER_DEBUG) || defined(AETHER_RELWITHDEB)
-		double divider = 1024. * 1024.;
+		float divider = 1024.f * 1024.f;
 		renderer_.end_draw_(evaluated_fps, heap::usage() / divider, lua_manager_.state_.memory_used() / divider);
 #else
 		renderer_.end_draw_();
